backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asn1crypto==1.5.1
attrs==25.1.0
bip-utils==2.9.3
bitarray==3.0.0
build==1.2.2.post1
CacheControl==0.14.2
cbor2==5.6.5
cdp-agentkit-core==0.0.11
cdp-langchain==0.0.13
cdp-sdk==0.15.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
ckzg==2.0.1
cleo==2.1.0
coincurve==20.0.0
crashtest==0.4.1
crcmod==1.7
cryptography==44.0.0
cytoolz==1.0.1
dataclasses-json==0.6.7
distlib==0.3.9
distro==1.9.0
dulwich==0.22.7
ecdsa==0.19.0
ed25519-blake2b==1.4.1
eth-account==0.13.4
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.1.0
eth-typing==5.1.0
eth-utils==5.2.0
eth_abi==5.2.0
fastjsonschema==2.21.1
filelock==3.17.0
frozenlist==1.5.0
h11==0.14.0
hexbytes==1.3.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
keyring==25.6.0
langchain==0.3.17
langchain-community==0.3.16
langchain-core==0.3.33
langchain-openai==0.2.14
langchain-text-splitters==0.3.5
langgraph==0.2.69
langgraph-checkpoint==2.0.10
langgraph-sdk==0.1.51
langsmith==0.3.4
marshmallow==3.26.1
more-itertools==10.6.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==2.2.2
openai==1.61.0
orjson==3.10.15
packaging==24.2
parsimonious==0.10.0
pkginfo==1.12.0
platformdirs==4.3.6
poetry==2.0.1
poetry-core==2.0.1
propcache==0.2.1
py-sr25519-bindings==0.2.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
PyJWT==2.10.1
PyNaCl==1.5.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyunormalize==16.0.0
PyYAML==6.0.2
RapidFuzz==3.12.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rlp==4.0.1
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
tenacity==9.0.0
tiktoken==0.8.0
tomlkit==0.13.2
toolz==1.0.0
tqdm==4.67.1
trove-classifiers==2025.1.15.22
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
virtualenv==20.29.1
web3==7.7.0
websockets==13.1
wheel==0.45.1
xattr==1.1.4
yarl==1.18.3
zstandard==0.23.0
